Example of Flax Issue
Is anyone having an issue with the player just floating up (along Y) continuously when the line:

pickActor.Position = Actor.Position;

is added in the Player.cs HandleMouseLook() function? If I remove that line, everything works as expected and matches the video. But if I add that line, the player and camera constantly increase their Y position (looks like the ground is shrinking away).